+++ src/globals.h
@@ -1,7 +1,7 @@
 #ifndef GLOBALS_H
 #define GLOBALS_H
 
-extern uint64_t NUM_THREADS;
+extern uint NUM_THREADS;
 extern std::string PROGRAM_NAME; 
 
 #endif // GLOBALS_H
